brainint: Add a jump cache
With this we only calculate the target jump position once for
each jump and cache the result. We even only do this lazyly so
if we would never jump on one particular loop we would also never
calculate the target jump position.

On my machine this reduces the runtime of the mandelbrot example
from 25sec to 15sec. ^^