brainbyte: Add clear byte code
A common pattern in brainfuck is a clear loop `[-]` which will set
the cell at the datapointer to zero.

Before this change we would have emited three Opcodes for this
loop whith a total of 20 bytes (brackets are especially expensive).

Now we only emit a single byte.

In the mandelbrot example we reduced the size of emited bytes
from 18001 to 15721 which is a reduction of 12%. Quite good. ^^

Unfortunatly I couldn't find any significant runtime improvements.