brainbyte: Implement postponing moves and disable them.
So the optimization should improve speed but doesn't because it
introduces new opcodes which decreases performance from ~7sec for
mandelbrot to ~14. I assume because more opcodes causes clang
to throw away a optimization causing in so bad performance.

So for now we will disable this optimization in brainbyte, however,
we will use it in the jit compilers so it stays in the code for now.