Implement dynasm implementation
I think...
Compiling seams to be a whole thing, but I have to go so that will
be a bliss of a task for tomorrow.

Anyway, I definitly understand how it works, which is already a
big step in my unresentles quest to understand jit compiler (and
the whole world of computing).

I am still happy to have a working jit implementation soon. ^^